refactor: clean up and shift `/parsers/parser.ts` -> `/parsers/index.ts`
