+++ src/infiniteLoops/__tests__/instrument.ts
@@ -1,7 +1,7 @@
 import { Program } from 'estree'
 
 import { mockContext } from '../../mocks/context'
-import { parse } from '../../parsers/parser'
+import { parse } from '../../parsers'
 import { Chapter } from '../../types'
 import { evaluateBinaryExpression, evaluateUnaryExpression } from '../../utils/operators'
 import {
