+++ src/runner/utils.ts
@@ -3,7 +3,7 @@ import { DebuggerStatement, Literal, Program } from 'estree'
 
 import { IOptions, Result } from '..'
 import { loadModuleTabs } from '../modules/moduleLoader'
-import { parseAt } from '../parsers/parser'
+import { parseAt } from '../parsers/utils'
 import { areBreakpointsSet } from '../stdlib/inspector'
 import { Context, Variant } from '../types'
 import { simple } from '../utils/walkers'
