+++ src/stepper/stepper.ts
@@ -2,7 +2,7 @@ import { generate } from 'astring'
 import * as es from 'estree'
 
 import * as errors from '../errors/errors'
-import { parse } from '../parsers/parser'
+import { parse } from '../parsers'
 import {
   BlockExpression,
   Context,
