+++ src/transpiler/__tests__/modules.ts
@@ -1,7 +1,7 @@
 import type { Identifier, Literal, MemberExpression, VariableDeclaration } from 'estree'
 
 import { mockContext } from '../../mocks/context'
-import { parse } from '../../parsers/parser'
+import { parse } from '../../parsers'
 import { Chapter } from '../../types'
 import { stripIndent } from '../../utils/formatters'
 import { transformImportDeclarations, transpile } from '../transpiler'
