+++ src/typeChecker/__tests__/source4Typed.test.ts
@@ -1,6 +1,6 @@
 import { parseError } from '../..'
 import { mockContext } from '../../mocks/context'
-import { parse } from '../../parsers/parser'
+import { parse } from '../../parsers'
 import { Chapter, Variant } from '../../types'
 
 let context = mockContext(Chapter.SOURCE_4, Variant.TYPED)
