+++ src/utils/testing.ts
@@ -5,7 +5,7 @@ import { transpileToGPU } from '../gpu/gpu'
 import { parseError, Result, runInContext } from '../index'
 import { transpileToLazy } from '../lazy/lazy'
 import { mockContext } from '../mocks/context'
-import { parse } from '../parsers/parser'
+import { parse } from '../parsers'
 import { transpile } from '../transpiler/transpiler'
 import { Chapter, Context, CustomBuiltIns, SourceError, Value, Variant } from '../types'
 import { stringify } from './stringify'
